tests: share 'file_data' fixture
